backend/api_gateway/requirements.txt
Django>=4.0
djangorestframework
kafka-python
psycopg2-binary
djangorestframework-simplejwt
django-cors-headers
python-decouple
django-extensions



